+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-17 14:46:52
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-05-17 18:55:50
+// Commit: fix(core): correct layout constraints and overflow in core
 
 class coreComponentState {
-  final String commitRef = '1223';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-05-17T14:46:52+05:00');
+  final String commitRef = '1226';
+  final String lastMessage = 'fix(core): correct layout constraints and overflow in core';
+  final DateTime timestamp = DateTime.parse('2026-05-17T18:55:50+05:00');
 }