+++ CHANGELOG.md
@@ -1227,3 +1227,4 @@
 - [2026-05-17 14:46] feat(core): add bi-directional event bus handler for core
 - [2026-05-17 15:14] fix(security): fix memory leak by cancelling security subscription
 - [2026-05-17 16:06] docs(i18n): update readme and API contracts for i18n
+- [2026-05-17 18:55] fix(core): correct layout constraints and overflow in core